Improve code quality and suppress HF token warnings
- Add scripts to download all models and suppress HF token warnings
- Fix Ruff linting errors in scripts and examples
- Add module docstrings to examples
- Update README with info about download_all_models script
- Fix Path.open() usage instead of open() in scripts
