+++ scripts/__init__.py
@@ -1,0 +1,3 @@
+"""Scripts package for TabPFN."""
+
+from __future__ import annotations
